User Controller Fixes and added game controller -MJ
