+++ GameHopper/Controllers/GameController.cs
@@ -1,0 +1,14 @@
+using GameHopper.Models;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.Build.Framework;
+
+namespace GameHopper;
+
+public class GameController : Controller {
+
+    [HttpGet]
+    public IActionResult Create()
+    {
+        return View();
+    }
+}
