+++ GameHopper/Models/User.cs
@@ -7,7 +7,7 @@ public abstract class User
     public string Name { get; set; }
     public int Id { get; set;}
 
-    public User(string name,int Id)
+        public User(string name,int Id)
 {
             Name = name;
     }
@@ -16,5 +16,26 @@ public abstract class User
             return Name;
     }
     }
+    public class GameMaster : User
+{
+    public ICollection<CreatedGames> CreatedGames { get; set; }
+    
+    public GameMaster(string name, int Id) : base(name, Id)
+    {
+        Name = name;
+    }
+
+    }
+
+public class CreatedGames
+{
+}
+
+public class Player : User
+{
+    public Player(string name, int Id) : base(name, Id)
+    {
+    }
+}
 }
